backend/requirements_auth.txt
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
PyJWT>=2.8.0
